+++ aws/eks/eks.ts
@@ -1,38 +1,18 @@
-import * as awsx from "@pulumi/awsx"; // Import AWS Crosswalk (awsx) resources for networking
 import * as eks from "@pulumi/eks"; // Import Pulumi EKS module for managing Amazon EKS clusters
 
-import { createRoles } from "./eks-iam"; // Import a function to create IAM roles from a custom module
+import { eksVpc } from "./networking"; // Import the VPC created for the EKS cluster
+import { instanceRoles } from "./eks-iam"; // Import the IAM roles for EKS worker nodes
 import {
   desiredSize, // Desired number of nodes in the node group
   eksClusterName, // Name of the EKS cluster
   eksNodeRootVolumeSize, // Size of the EBS root volume attached to the EKS nodes
+  eksPublicAccess, // CIDR blocks for public access to the EKS API
   eksVersion, // EKS Kubernetes version
-  eksVPCCIDRBlock, // CIDR block for the VPC
   instanceType, // EC2 instance type for the worker nodes
   minSize, // Minimum number of worker nodes
   maxSize, // Maximum number of worker nodes
   tags, // Tags to attach to AWS resources
 } from "./variables"; // Import variables from the variables file
-
-// Create a VPC for our EKS cluster with public and private subnets
-export const eksVpc = new awsx.ec2.Vpc(`${eksClusterName}-vpc`, {
-  assignGeneratedIpv6CidrBlock: false, // Disable IPv6 CIDR block assignment
-  cidrBlock: eksVPCCIDRBlock, // Use the specified CIDR block
-  enableDnsSupport: true, // Enable DNS support in the VPC
-  enableNetworkAddressUsageMetrics: true, // Enable network address usage metrics
-  subnetStrategy: "Auto", // Automatically create subnets
-  subnetSpecs: [
-    { type: awsx.ec2.SubnetType.Public }, // Public subnets for external access
-    { type: awsx.ec2.SubnetType.Private }, // Private subnets for internal resources
-  ],
-  tags: tags, // Attach tags to the VPC and associated subnets
-});
-
-// Create IAM roles for EKS worker nodes
-export const instanceRoles = createRoles(
-  `${eksClusterName}-instance-role`, // Base name for the roles
-  3 // Create 3 instance roles for the EKS node group
-);
 
 // Create an EKS cluster with custom configurations
 export const cluster = new eks.Cluster(`${eksClusterName}-cluster`, {
@@ -45,6 +25,8 @@ export const cluster = new eks.Cluster(`${eksClusterName}-cluster`, {
     "scheduler",
   ], // Enable cluster logging for key components
   endpointPublicAccess: true, // Enable public access to the EKS API endpoint
+  endpointPrivateAccess: true, // Enable public access to the EKS API endpoint
+  publicAccessCidrs: eksPublicAccess, // List of CIDR blocks for public access to the EKS API
   instanceRoles: instanceRoles, // Attach the created instance roles to the cluster
   name: eksClusterName, // Name of the EKS cluster
   nodeAssociatePublicIpAddress: false, // Disable public IP addresses for worker nodes
